seq: fixed a corner case in ASYNC put/get
If epicsEventTryWait succeeds, there could still be a previous request
pending that has timed out. This happens e.g. if the user code did not
call pvGetComplete resp. pvPutComplete. In that case we must reset the
request pointer to NULL in order to invalidate the request so it gets
ignored whenever it finally arrives.