docs: branch name needs a dash to work for darcs get
Installation.txt says you can check out the darcs repository like this:

  darcs get http://www-csr.bessy.de/control/SoftDist/sequencer/repo/<branch>

It then lists the branches, which to me should be the exact string I
need to use in place of "<branch>" to construct a valid URL for the
repository I want, but it's not.  Fix this by replacing the space in the
listed branch names with a dash.