fix: must not take program lock around pvVarDestroy [amended for branch-2-1]
From EPICS base 3.14.12.3 to 3.14.12.4 there was a change (a bug fix) in the
ca_clear_channel implementation with the effect that it waits for any
already active callback to complete. This could cause a deadlock between CA
callbacks and a state set thread calling pvAssign because both need to take
the program instance mutex lock. The solution is to release the lock before
pvVarDestroy is called (which in turn calls ca_clear_channel) and to re-take
it afterwards.

This patch also fixes a few places in the code where the dbch member of a
channel was stored inside a local variable before the program lock is taken.
This was done to ensure that concurrent writes to parts of the dbch
structure do not collide. Note that we still assume that (aligned) pointer
read and write operations are atomic w.r.t. to each other w/o taking a mutex
lock.