seq: disabled 'no pending put request' error message in pvGet/PutComplete
As Janez pointed out, it makes sense to call pvPutComplete for a channel
array in which only some of the channels have pending puts. This resulted in
lots of annoying messages on the console. Note that calling pvPutComplete
with no pending put is completely harmless (it just returns TRUE).

For symmetry, I have disabled the corresponding message in pvGetComplete,
too, even though the latter does not handle channel arrays.